move const
