+++ web/src/api/api_helper.ts
@@ -1,3 +1,6 @@
+export const API_BASE_URL =
+  import.meta.env?.VITE_API_BASE_URL || "http://localhost:1323/api/v1";
+
 export function getAuthHeader(): HeadersInit {
   const token = localStorage.getItem("token");
   if (!token) {
