--- web/src/config.ts
@@ -1,2 +1,0 @@
-export const API_BASE_URL =
-  import.meta.env?.VITE_API_BASE_URL || "http://localhost:1323/api/v1";
